Validaciones
Se hicieron ajustes para validar el el modulo de asignacion de
apoyos, en los diferentes archivos

+++ app/database/migrations/2015_07_16_131703_create_apoyos.php
@@ -15,7 +15,7 @@ class CreateApoyos extends Migration {
 		Schema::create('apoyos', function(Blueprint $table)
 		{
 			$table->increments('id');
-			$table->float('monto');
+			$table->double('monto',7,4);
 			$table->dateTime('fecha');
 			$table->string('concepto', 1024);
 			$table->enum('periodicidad', array('Un solo pago','Quincenal', 'Mensual','Bimestral,','Semestral','Anual','Otro'));
